Add base Attraction model
Add base model to hold attributes shared between Destinations and Events.
Refactor Destination to inherit from Attraction.
