+++ app/src/main/java/com/gophillygo/app/data/GpgDatabase.java
@@ -6,7 +6,7 @@ import android.arch.persistence.room.TypeConverters;
 
 import com.gophillygo.app.data.models.Destination;
 
-@Database(version=4, entities={Destination.class})
+@Database(version=5, entities={Destination.class})
 @TypeConverters({RoomConverters.class})
 public abstract class GpgDatabase extends RoomDatabase {
     abstract public DestinationDao destinationDao();
