+++ app/src/main/java/com/gophillygo/app/data/models/Destination.java
@@ -3,7 +3,6 @@ package com.gophillygo.app.data.models;
 import android.arch.persistence.room.ColumnInfo;
 import android.arch.persistence.room.Embedded;
 import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.PrimaryKey;
 
 import com.google.gson.annotations.SerializedName;
 
@@ -12,7 +11,7 @@ import java.util.ArrayList;
 
 
 @Entity
-public class Destination {
+public class Destination extends Attraction {
 
     private static final NumberFormat numberFormatter = NumberFormat.getNumberInstance();
     static {
@@ -20,49 +19,23 @@ public class Destination {
         numberFormatter.setMaximumFractionDigits(2);
     }
 
-    @PrimaryKey
-    private final int id;
-
-    private final int placeID;
-    private final String name;
-    private final boolean accessible;
-    private final String image;
     private final String city;
-    private final boolean cycling;
-    private final String description;
-    private final int priority;
     private final String state;
     private final String address;
-    private final ArrayList<String> activities;
 
     @Embedded
     private final DestinationLocation location;
-
-    @Embedded
-    private final DestinationAttributes attributes;
 
     @ColumnInfo(name = "watershed_alliance")
     @SerializedName("watershed_alliance")
     private final boolean watershedAlliance;
 
-    @ColumnInfo(name = "website_url")
-    @SerializedName("website_url")
-    private final String websiteUrl;
-
-    @ColumnInfo(name = "wide_image")
-    @SerializedName("wide_image")
-    private final String wideImage;
-
-    @ColumnInfo(name = "is_event")
-    @SerializedName("is_event")
-    private final boolean isEvent;
+    @Embedded
+    private final DestinationAttributes attributes;
 
     @ColumnInfo(name = "zipcode")
     @SerializedName("zipcode")
     private final String zipCode;
-
-    // timestamp is not final, as it is set on database save, and not by serializer
-    private long timestamp;
 
     // convenience property to track distance to each destination
     private float distance;
@@ -73,36 +46,19 @@ public class Destination {
                        int priority, String state, String address, DestinationLocation location,
                        DestinationAttributes attributes, boolean watershedAlliance, String websiteUrl,
                        String wideImage, boolean isEvent, ArrayList<String> activities) {
-        this.id = id;
-        this.placeID = placeID;
-        this.name = name;
-        this.accessible = accessible;
-        this.image = image;
+
+        // initialize Attraction
+        super(id, placeID, name, accessible, image, cycling, description, priority, websiteUrl,
+                wideImage, isEvent, activities);
+
         this.city = city;
-        this.cycling = cycling;
         this.zipCode = zipCode;
-        this.description = description;
-        this.priority = priority;
         this.state = state;
         this.address = address;
         this.watershedAlliance = watershedAlliance;
-        this.websiteUrl = websiteUrl;
-        this.wideImage = wideImage;
-        this.isEvent = isEvent;
 
         this.location = location;
         this.attributes = attributes;
-        this.activities = activities;
-    }
-
-    /**
-     * Timestamp entries. Timestamp value does not come from query result; it should be set
-     * on database save. Gson serializer will initialize value to zero.
-     *
-     * @param timestamp Time in milliseconds since Unix epoch
-     */
-    public void setTimestamp(long timestamp) {
-        this.timestamp = timestamp;
     }
 
     public void setDistance(float distance) {
@@ -115,40 +71,8 @@ public class Destination {
         this.formattedDistance = formattedDistance;
     }
 
-    public int getId() {
-        return id;
-    }
-
-    public int getPlaceID() {
-        return placeID;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public boolean isAccessible() {
-        return accessible;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
     public String getCity() {
         return city;
-    }
-
-    public boolean isCycling() {
-        return cycling;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public int getPriority() {
-        return priority;
     }
 
     public String getState() {
@@ -167,28 +91,8 @@ public class Destination {
         return attributes;
     }
 
-    public boolean isWatershedAlliance() {
-        return watershedAlliance;
-    }
-
-    public String getWebsiteUrl() {
-        return websiteUrl;
-    }
-
-    public String getWideImage() {
-        return wideImage;
-    }
-
-    public boolean isEvent() {
-        return isEvent;
-    }
-
     public String getZipCode() {
         return zipCode;
-    }
-
-    public long getTimestamp() {
-        return timestamp;
     }
 
     public float getDistance() {
@@ -199,7 +103,7 @@ public class Destination {
         return formattedDistance;
     }
 
-    public ArrayList<String> getActivities() {
-        return activities;
+    public boolean isWatershedAlliance() {
+        return watershedAlliance;
     }
 }
